refactor(chat): symmetric tool executors, shared asset-privacy module, honest invariant comments
- read_notes executor moves to read-notes.ts mirroring read-assets.ts;
  tools.ts is now registration + names + unions only, and the re-export
  facade is gone (each symbol has one home).
- classifyAsset/classifyAssetFromNotes/AssetVerdict move to
  actions/asset-privacy.ts — the verdict is shared by the generation pass
  and the chat read path, not a description-file helper.
- The 'never AI-reachable' comments in indexing and Plan 20 now say what
  is actually true post-read_assets: note text and previews are never
  widened; chat reaches descriptions only via the gated tool.
