+++ packages/core/src/ai/chat/tools.test.ts
@@ -5,20 +5,24 @@ import type { DailyNoteRow, DailyNotesRange } from '../../indexing/queries'
 import type { RecentNoteRow, RecentNotesOptions } from '../../indexing/note-list'
 import {
   ASSET_UNAVAILABLE_ERROR,
-  buildNoteTools,
-  INVALID_TAG_ERROR,
   MAX_ASSET_DESCRIPTION_CHARS,
-  MAX_DAILY_NOTE_DAYS,
-  MAX_NOTE_CONTENT_CHARS,
   NO_ASSET_DESCRIPTION_ERROR,
   NOT_AN_ASSET_ERROR,
+  type ReadAssetResult,
+  type ReadAssetsOutput,
+} from './read-assets'
+import {
+  MAX_NOTE_CONTENT_CHARS,
+  type ReadNoteResult,
+  type ReadNotesOutput,
+} from './read-notes'
+import {
+  buildNoteTools,
+  INVALID_TAG_ERROR,
+  MAX_DAILY_NOTE_DAYS,
   type ListDailyNotesOutput,
   type ListRecentNotesOutput,
   type NoteTools,
-  type ReadAssetResult,
-  type ReadAssetsOutput,
-  type ReadNoteResult,
-  type ReadNotesOutput,
   type SearchNotesOutput,
 } from './tools'
